compose: move compose-related classes into compose package
