+++ k9mail/src/main/java/com/fsck/k9/activity/MessageCompose.java
@@ -64,8 +64,11 @@ import com.fsck.k9.Identity;
 import com.fsck.k9.K9;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.R;
-import com.fsck.k9.activity.CryptoSettingsDialog.OnCryptoModeChangedListener;
-import com.fsck.k9.activity.RecipientPresenter.CryptoMode;
+import com.fsck.k9.activity.compose.ComposeCryptoStatus;
+import com.fsck.k9.activity.compose.CryptoSettingsDialog.OnCryptoModeChangedListener;
+import com.fsck.k9.activity.compose.RecipientMvpView;
+import com.fsck.k9.activity.compose.RecipientPresenter;
+import com.fsck.k9.activity.compose.RecipientPresenter.CryptoMode;
 import com.fsck.k9.activity.loader.AttachmentContentLoader;
 import com.fsck.k9.activity.loader.AttachmentInfoLoader;
 import com.fsck.k9.activity.misc.Attachment;
@@ -94,6 +97,7 @@ import com.fsck.k9.message.IdentityField;
 import com.fsck.k9.message.IdentityHeaderParser;
 import com.fsck.k9.message.InsertableHtmlContent;
 import com.fsck.k9.message.MessageBuilder;
+import com.fsck.k9.message.PgpMessageBuilder;
 import com.fsck.k9.message.QuotedTextMode;
 import com.fsck.k9.message.SimpleMessageBuilder;
 import com.fsck.k9.message.SimpleMessageFormat;
