+++ k9mail/src/main/java/com/fsck/k9/activity/compose/RecipientLoader.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.activity;
+package com.fsck.k9.activity.compose;
 
 
 import java.util.ArrayList;
