+++ k9mail/src/main/java/com/fsck/k9/view/RecipientSelectView.java
@@ -36,8 +36,8 @@ import com.fsck.k9.K9;
 import com.fsck.k9.R;
 import com.fsck.k9.activity.AlternateRecipientAdapter;
 import com.fsck.k9.activity.AlternateRecipientAdapter.AlternateRecipientListener;
-import com.fsck.k9.activity.RecipientAdapter;
-import com.fsck.k9.activity.RecipientLoader;
+import com.fsck.k9.activity.compose.RecipientAdapter;
+import com.fsck.k9.activity.compose.RecipientLoader;
 import com.fsck.k9.mail.Address;
 import com.fsck.k9.view.RecipientSelectView.Recipient;
 import com.tokenautocomplete.TokenCompleteTextView;
